chore: rename components folders to modules
