+++ www/pages/contacts/modules/index.js
@@ -1,0 +1,1 @@
+export { default as ProjectInfo } from './react-project-info';
