+++ www/pages/contacts/modules/react-project-info/index.js
@@ -1,0 +1,1 @@
+export { default } from './ProjectInfo';
