+++ www/pages/error/ErrorPage.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import ErrorPage from './ErrorPage';
-import { AppTree } from '../../shared/test-utils/components';
+import { AppTree } from '../../shared/test-utils/modules';
 
 it('should render internal server error when `statusCode` is not 404', () => {
     const { getByText } = render(
