--- www/shared/components/index.js
@@ -1,4 +1,0 @@
-export { default as CookieBanner } from './cookie-banner';
-export { default as Footer } from './footer';
-export { default as Header } from './header';
-export { default as MainLayout } from './main-layout';
