+++ www/shared/modules/index.js
@@ -1,0 +1,4 @@
+export { default as CookieBanner } from './react-cookie-banner';
+export { default as Footer } from './react-footer';
+export { default as Header } from './react-header';
+export { default as MainLayout } from './react-main-layout';
