+++ www/shared/modules/react-cookie-banner/CookieBanner.js
