+++ www/shared/modules/react-cookie-banner/index.js
