+++ www/shared/modules/react-footer/Footer.js
