+++ www/shared/modules/react-footer/Footer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 import Footer from './Footer';
-import { AppTree } from '../../test-utils/components';
+import { AppTree } from '../../test-utils/modules';
 
 it('should render correctly', () => {
     const { container } = render((
