+++ www/shared/modules/react-footer/index.js
