--- www/shared/test-utils/components/index.js
@@ -1,1 +1,0 @@
-export { default as AppTree, AppTreeWrapper } from './app-tree';
