+++ www/shared/test-utils/modules/index.js
@@ -1,0 +1,1 @@
+export { default as AppTree, AppTreeWrapper } from './react-app-tree';
